# enhance: Optimize save collection target latency (#38345)